Ich hatte einen Haufen falscher Fehlermeldungen.

+++ src/main/java/datastorage/PatientDAO.java
@@ -2,6 +2,7 @@ package datastorage;
 
 import model.Patient;
 import utils.DateConverter;
+
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -38,7 +39,7 @@ public class PatientDAO extends DAOimp<Patient> {
      * @return <code>String</code> with the generated SQL.
      */
     @Override
-    protected String getReadByIDStatementString(long key) {
+    protected String getReadByIDStatementString(int key) {
         return String.format("SELECT * FROM patient WHERE pid = %d", key);
     }
 
@@ -103,7 +104,7 @@ public class PatientDAO extends DAOimp<Patient> {
      * @return <code>String</code> with the generated SQL.
      */
     @Override
-    protected String getDeleteStatementString(long key) {
+    protected String getDeleteStatementString(int key) {
         return String.format("Delete FROM patient WHERE pid=%d", key);
     }
 }
