+++ src/main/java/datastorage/TreatmentDAO.java
@@ -2,6 +2,7 @@ package datastorage;
 
 import model.Treatment;
 import utils.DateConverter;
+
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -26,7 +27,7 @@ public class TreatmentDAO extends DAOimp<Treatment> {
     }
 
     @Override
-    protected String getReadByIDStatementString(long key) {
+    protected String getReadByIDStatementString(int key) {
         return String.format("SELECT * FROM treatment WHERE tid = %d", key);
     }
 
@@ -69,7 +70,7 @@ public class TreatmentDAO extends DAOimp<Treatment> {
     }
 
     @Override
-    protected String getDeleteStatementString(long key) {
+    protected String getDeleteStatementString(int key) {
         return String.format("Delete FROM treatment WHERE tid= %d", key);
     }
 
@@ -86,7 +87,7 @@ public class TreatmentDAO extends DAOimp<Treatment> {
         return String.format("SELECT * FROM treatment WHERE pid = %d", pid);
     }
 
-    public void deleteByPid(long key) throws SQLException {
+    public void deleteByPid(int key) throws SQLException {
         Statement st = conn.createStatement();
         st.executeUpdate(String.format("Delete FROM treatment WHERE pid= %d", key));
     }
